developer

+++ src/Model/Sala.php
@@ -25,9 +25,18 @@ class Sala
     {
         try {
             $this->validateToken($token);
-            $db = new Connection();
-            $sql = 'SELECT * FROM sala';
-            return $db->query($sql);
+            $decoded = $this->validateToken($token);
+            if ($decoded->perfil == 'administrador_supremo' || $decoded->perfil == 'administrador') {
+                $db = new Connection();
+                $sql = 'SELECT * FROM sala';
+                if ($salas = $db->query($sql)) {
+                    return $salas;
+                } else {
+                    throw new Exception('Nenhuma sala encontrada');
+                }
+            } else {
+                throw new Exception('Acesso negado');
+            }
         } catch (Exception $e) {
             return $e->getMessage();
         }
@@ -37,51 +46,83 @@ class Sala
     {
         try {
             $this->validateToken($token);
-            $db = new Connection();
-            $sql = 'SELECT * FROM sala WHERE id_sala = :id';
-            return $db->query($sql, ['id' => $id]);
+            $decoded = $this->validateToken($token);
+            if ($decoded->perfil == 'administrador_supremo' || $decoded->perfil == 'administrador') {
+                $db = new Connection();
+                $sql = 'SELECT * FROM sala WHERE id_sala = :id';
+                try {
+                    if ($sala = $db->query($sql, ['id' => $id])) {
+                        return $sala;
+                    } else {
+                        throw new Exception('Sala não encontrada');
+                    }
+                } catch (Exception $e) {
+                    return $e->getMessage();
+                }
+            } else {
+                throw new Exception('Acesso negado');
+            }
         } catch (Exception $e) {
             return $e->getMessage();
         }
     }
 
-    function cadastrar($id_sala, $numero_sala, $capacidade_sala, $id_equipamento, $token)
+    function cadastrar($nome_sala, $numero_sala, $capacidade_sala, $id_equipamento, $token)
     {
         try {
             $this->validateToken($token);
-            $db = new Connection();
-            $sql = 'INSERT INTO sala (
-                id_sala, numero_sala, capacidade_sala, id_equipamento
-            ) VALUES (
-                :id_sala, :numero_sala, :capacidade_sala, :id_equipamento
-            )';
-            return $db->query_insert($sql, [
-                'id_sala' => $id_sala,
-                'numero_sala' => $numero_sala,
-                'capacidade_sala' => $capacidade_sala,
-                'id_equipamento' => $id_equipamento
-            ]);
+            $decoded = $this->validateToken($token);
+            if ($decoded->perfil == 'administrador_supremo' || $decoded->perfil == 'administrador') {
+                try{
+                    $db = new Connection();
+                    $sql = 'INSERT INTO sala (nome_sala, numero_sala, capacidade_sala, id_equipamento) VALUES (:nome_sala, :numero_sala, :capacidade_sala, :id_equipamento)';
+                    if ($db->query_insert($sql, [
+                        'nome_sala' => $nome_sala,
+                        'numero_sala' => $numero_sala,
+                        'capacidade_sala' => $capacidade_sala,
+                        'id_equipamento' => $id_equipamento
+                    ])) {
+                        return ['success' => 'Sala cadastrada com sucesso'];
+                    } else {
+                        throw new Exception('Erro ao cadastrar sala');
+                    }
+                } catch (Exception $e) {
+                    return $e->getMessage();
+                }
+            } else {
+                throw new Exception('Acesso negado');
+            }
         } catch (Exception $e) {
             return $e->getMessage();
         }
     }
 
-    function atualizar($id, $numero_sala, $capacidade_sala, $id_equipamento, $token)
+    function atualizar($id, $nome_sala, $numero_sala, $capacidade_sala, $id_equipamento, $token)
 {
     try {
         $this->validateToken($token);
-        $db = new Connection();
-        $sql = 'UPDATE sala SET 
-            numero_sala = :numero_sala, 
-            capacidade_sala = :capacidade_sala,
-            id_equipamento = :id_equipamento
-            WHERE id_sala = :id_sala';
-        return $db->query_update($sql, [
-            'id_sala' => $id, 
-            'numero_sala' => $numero_sala,
-            'capacidade_sala' => $capacidade_sala,
-            'id_equipamento' => $id_equipamento
-        ]);
+        $decoded = $this->validateToken($token);
+        if ($decoded->perfil == 'administrador_supremo' || $decoded->perfil == 'administrador') {
+            try{
+                $db = new Connection();
+                $sql = 'UPDATE sala SET nome_sala = :nome_sala, numero_sala = :numero_sala, capacidade_sala = :capacidade_sala, id_equipamento = :id_equipamento WHERE id_sala = :id';
+                if ($db->query_update($sql, [
+                    'id' => $id,
+                    'nome_sala' => $nome_sala,
+                    'numero_sala' => $numero_sala,
+                    'capacidade_sala' => $capacidade_sala,
+                    'id_equipamento' => $id_equipamento
+                ])) {
+                    return ['success' => 'Sala atualizada com sucesso'];
+                } else {
+                    throw new Exception('Erro ao atualizar sala');
+                }
+            } catch (Exception $e) {
+                return $e->getMessage();
+            }
+        } else {
+            throw new Exception('Acesso negado');
+        }
     } catch (Exception $e) {
         return $e->getMessage();
     }
@@ -92,9 +133,18 @@ class Sala
     {
         try {
             $this->validateToken($token);
-            $db = new Connection();
-            $sql = 'DELETE FROM sala WHERE id_sala = :id';
-            return $db->query_delete($sql, ['id' => $id]);
+            $decoded = $this->validateToken($token);
+            if ($decoded->perfil == 'administrador_supremo' || $decoded->perfil == 'administrador') {
+                $db = new Connection();
+                $sql = 'DELETE FROM sala WHERE id_sala = :id';
+                if ($db->query_delete($sql, ['id' => $id])) {
+                    return ['success' => 'Sala excluída com sucesso'];
+                } else {
+                    throw new Exception('Erro ao excluir sala');
+                }
+            } else {
+                throw new Exception('Acesso negado');
+            }
         } catch (Exception $e) {
             return $e->getMessage();
         }
